👌 Minor cleanups
