+++ upgrades/0.0.97.seperate-ns-from-app/pkg/upgrade/argocd.go
@@ -1,5 +1,12 @@
 package upgrade
 
+import (
+	"fmt"
+
+	"github.com/spf13/afero"
+	"sigs.k8s.io/yaml"
+)
+
 const (
 	argoCDApplicationAPIVersion = "argoproj.io/v1alpha1"
 	argoCDApplicationKind       = "Application"
@@ -20,4 +27,29 @@ func (receiver argoCDApplication) Valid() bool {
 	}
 
 	return true
+}
+
+func isPathAnArgoCDApplication(fs *afero.Afero, absoluteApplicationPath string) (bool, error) {
+	stat, err := fs.Stat(absoluteApplicationPath)
+	if err != nil {
+		return false, fmt.Errorf("stating file: %w", err)
+	}
+
+	if stat.IsDir() {
+		return false, nil
+	}
+
+	rawFile, err := fs.ReadFile(absoluteApplicationPath)
+	if err != nil {
+		return false, fmt.Errorf("reading file: %w", err)
+	}
+
+	var potentialApp argoCDApplication
+
+	err = yaml.Unmarshal(rawFile, &potentialApp)
+	if err != nil {
+		return false, fmt.Errorf("unmarshalling: %w", err)
+	}
+
+	return potentialApp.Valid(), nil
 }
