+++ upgrades/0.0.97.seperate-ns-from-app/pkg/upgrade/fs.go
@@ -1,0 +1,31 @@
+package upgrade
+
+import (
+	"fmt"
+	"path"
+	"strings"
+
+	"github.com/spf13/afero"
+)
+
+func copyFile(fs *afero.Afero, sourcePath string, destinationPath string) error {
+	content, err := fs.Open(sourcePath)
+	if err != nil {
+		return fmt.Errorf("reading: %w", err)
+	}
+
+	defer func() {
+		_ = content.Close()
+	}()
+
+	err = fs.WriteReader(destinationPath, content)
+	if err != nil {
+		return fmt.Errorf("writing: %w", err)
+	}
+
+	return nil
+}
+
+func filenameWithoutExtension(filename string) string {
+	return strings.Replace(path.Base(filename), path.Ext(filename), "", 1)
+}
